Heap allocate log details
We cannot predict the size of log and we don't want to crash the
program by trying to allocate too much memory on the stack.
Also declared Log as a typedef for ease of programming.

Created another 'Async' function which still doesn't behave how I want
it to, we'll eventually get there :(